Make group window large
The arriving record from either the x or y streams appears to identify the
tumbling window into which the joined record is placed.

If the sliding window for a join crosses the boundary of two adjacent tumbling aggregation windows,
then, an arriving y record can either fall into either of the tumbling
windows. This means that there is more that one aggregation set for y
records joined to the same x record!